Add files via upload
Assignment (Capstone Project)
MediBuddy is a digital healthcare platform for an inpatient hospitalization, outpatient services, and corporate wellness benefits. MediBuddy is the award-winning technology platform from Medi Assist that transforms the health insurance industry at the very core. MediBuddy, with its diverse offerings for various stakeholders of the health benefits industry, makes the process of discovering, accessing, utilizing, and monitoring health benefits seamless, real-time, and virtually paperless. The company was founded in 2000 and based in Bangalore, Karnataka, India.
In the assignment below, there are two datasets, one having the age and BMI(Body Mass Index) that is received from a pre health checkup before the policy cover; and the other dataset includes some personal details such as the no. of children, whether the person smokes or not, the location where the person is from etc.
Share your analysis for the below:
1. Does the gender of the person matter for the company as a constraint for extending policies?
2. What is the average amount of money the company spent over each policy cover?
3. Could you advice if the company needs to offer separate policies based upon the geographic location of the person?
4. Does the no. of dependents make a difference in the amount claimed?
5. Does a study of persons BMI get the company any idea for the insurance claim that it would extend?
6. Is it needed for the company to understand whether the person covered is a smoker or a non-smoker?
7. Does age have any barrier on the insurance claimed?
8. Can the company extend certain discounts after checking the health status (BMI) in this case?
In addition to the above analysis, build a machine learning model to predict the amount spent to the utmost accuracy. Validate and hyper parametrically tune the model to get the best results.